Moved vertex color responsibilities out into their own classes
The QuadMesh has enough to do without worrying about vertex colors, too.  Moved that responsibility into two separate classes.  TerrainColorizer colors the basic terrain and EdgeColorizer is a decorator that adds coloring of the mesh boundaries.  The specs for EdgeColorizer have kind of a funky structure because MSpec doesn't have any data-driven facilities at the moment.  I may revisit that to see if there's a better way to write those.