Replaced the dynamic clipping plane system with a dynamic scaling system
The existing system to dynamically set the clipping planes for each frame worked ok but I didn't like the way the code was designed, and it was complicating the effort to add view frustum draw filtering.  I removed that system and replaced it with a dynamic draw scaling system that guarentees that all objects are rendered inside our (currently fixed) clipping planes.  There is significantly more z-fighting with this system alone but later we can implement a better system for dynamically moving the clipping planes as well and have the best of both worlds.