Fixed a failing StructureMap unit test
I had changed QuadMeshRenderer to take a BasicEffect in the constructor but that was messing up StructureMap in the unit tests because it can't create a BasicEffect at unit test time.  It doesn't need to be created by the container anyway (it has a factory to do that) so I excluded it from the assembly scan.